Move DetectorStateData to the stateful detector handler file.
